requirements.txt
asgiref~=3.8.1
attrs~=23.1.0
certifi~=2023.7.22
cffi~=1.16.0
charset-normalizer~=3.3.2
click~=8.1.7
click-plugins~=1.1.1
cligj~=0.7.2
colorama~=0.4.6
contourpy~=1.2.0
cycler~=0.12.1
Django~=5.0.4
et-xmlfile~=1.1.0
fiona~=1.9.5
outcome~=1.3.0.post0
pyproj~=3.6.1
python-dotenv~=1.0.1
setuptools~=68.0.0
sqlparse~=0.4.4
trio~=0.23.1
trio-websocket~=0.11.1
tzdata~=2024.1
wheel~=0.41.2
wsproto~=1.2.0
psycopg2~=2.9.9
djangorestframework~=3.15.1